perf: remove unnecessary rpc requests
